Refactored About components and Login Page
