+++ src/features/About/data/People.tsx
@@ -1,0 +1,33 @@
+import eric from "../../../assets/images/about/eric.jpg"
+import maxwell from "../../../assets/images/about/maxwell.jpg";
+import ramiro from "../../../assets/images/about/ramiro.jpg";
+import yogi from "../../../assets/images/about/yogi.jpg";
+
+const people = [
+    {
+      img: ramiro,
+      name: "Ramiro",
+      position: "Software Developer",
+      linkedIn: "https://www.linkedin.com/in/ramiro-pena/",
+    },
+    {
+      img: yogi,
+      name: "Yogi",
+      position: "Software Developer",
+      linkedIn: "https://www.linkedin.com/in/yogi-zipori/",
+    },
+    {
+      img: maxwell,
+      name: "Max",
+      position: "Software Developer",
+      linkedIn: "https://www.linkedin.com/in/maxwell-morrow-6a3085228/",
+    },
+    {
+      img: eric,
+      name: "Eric",
+      position: "Software Developer",
+      linkedIn: "https://www.linkedin.com/in/hoov/",
+    },
+  ];
+
+  export default people;
