+++ src/features/login/LoginHeader.tsx
@@ -2,10 +2,10 @@ const LoginHeader = () => {
 
     return (
         <header>
-            <h5 className="card-title font-weight-bold mb-5 fs-5">
+            <h5 className="font-weight-bold ">
                 CAREER PLAN TODAY
             </h5>
-            <h5 className="card-title mb-5 fw-light fs-5">
+            <h5 className="mb-5 fw-light fs-5">
                 Sign in to your account
             </h5>
         </header>
