+++ src/pages/AboutPage.tsx
@@ -1,0 +1,11 @@
+import AboutPeople from "../features/About/AboutPeople";
+
+const AboutPage = () => {
+    return (
+        <div>
+            <AboutPeople />
+        </div>
+    )
+}
+
+export default AboutPage;
